Sign Up Form Validation
Changed the sign up process of the signup.js file making it easier to understand how user input is collected and verified. Said verification is now implemented, with the current verification only being character length for now. If the boundaries of the verification are not met, the sign up form is not submitted until it is corrected. Also installed a new npm download, 'npm-hook-form' for client.