Button Displays and Username Display
Buttons show properly based on the user log in state, but the username display is not working properly. It's displaying a welcome message before logging in, and once logged in, the username that is being passed is not being shown. I believe the error is located "useEffect" in App.js, because my trip for displaying the passed username is not being displayed.